editoast: refactor Identifier in ModelV2 derive macro
