+++ editoast/editoast_derive/src/modelv2.rs
@@ -1,3 +1,5 @@
+mod identifier;
+
 use std::{
     collections::{HashMap, HashSet},
     ops::{Deref, DerefMut},
@@ -12,33 +14,13 @@ use proc_macro2::{Ident, Span, TokenStream};
 use quote::quote;
 use syn::DeriveInput;
 
-#[derive(Debug, Clone, PartialEq, Hash, Eq)]
-enum Identifier {
-    Field(syn::Ident),
-    Compound(Vec<syn::Ident>),
-}
+use identifier::Identifier;
 
 impl Identifier {
     fn get_field<'a>(&self, config: &'a ModelConfig) -> Option<&'a ModelField> {
         match self {
             Self::Field(ident) => config.fields.get(ident),
             Self::Compound(_) => None,
-        }
-    }
-
-    fn get_idents(&self) -> Vec<syn::Ident> {
-        match self {
-            Self::Field(ident) => vec![ident.clone()],
-            Self::Compound(idents) => idents.clone(),
-        }
-    }
-
-    fn get_ident_lvalue(&self) -> TokenStream {
-        match self {
-            Self::Field(ident) => quote! { #ident },
-            Self::Compound(idents) => {
-                quote! { (#(#idents),*) }
-            }
         }
     }
 
@@ -54,43 +36,6 @@ impl Identifier {
                     .map(|ident| &config.fields.get(ident).unwrap().ty);
                 quote! { (#(#ty),*) }
             }
-        }
-    }
-}
-
-fn extract_ident_of_path(path: &syn::Path) -> Result<syn::Ident> {
-    let mut segments = path.segments.iter();
-    let first = segments.next().unwrap();
-    if segments.next().is_some() {
-        Err(darling::Error::custom(
-            "Model: unexpected path in 'identifier' expression",
-        ))
-    } else {
-        Ok(first.ident.clone())
-    }
-}
-
-impl darling::FromMeta for Identifier {
-    fn from_expr(expr: &syn::Expr) -> Result<Self> {
-        match expr {
-            syn::Expr::Path(path) => Ok(Identifier::Field(extract_ident_of_path(&path.path)?)),
-            syn::Expr::Tuple(tuple) => {
-                let mut idents = Vec::new();
-                for expr in tuple.elems.iter() {
-                    match expr {
-                        syn::Expr::Path(path) => {
-                            idents.push(extract_ident_of_path(&path.path)?);
-                        }
-                        _ => return Err(darling::Error::custom(
-                            "Model: invalid compound 'identifier' expression: must be a tuple of idents",
-                        )),
-                    }
-                }
-                Ok(Identifier::Compound(idents))
-            }
-            _ => Err(darling::Error::custom(
-                "Model: invalid 'identifier' expression",
-            )),
         }
     }
 }
